requirements.txt
openai>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
pdfplumber>=0.11.0
chromadb>=0.5.0
